+++ packages/sampler/stories/qa-stories/Header.js
@@ -1,10 +1,9 @@
 import {Header} from '@enact/moonstone/Panels';
 import React from 'react';
 import {storiesOf} from '@kadira/storybook';
-import {withKnobs, text} from '@kadira/storybook-addon-knobs';
+import {text} from '@kadira/storybook-addon-knobs';
 
 storiesOf('Header')
-	.addDecorator(withKnobs)
 	.addWithInfo(
 		'with RTL text',
 		() => (