+++ packages/sampler/CHANGELOG.md
@@ -2,6 +2,16 @@
 
 The following is a curated list of changes in the Enact Sampler, newest changes on the top.
 
+## [unreleased]
+
+## Fixed
+
+- `largeText` and `locale` knobs to not disappear after use
+
+## [1.1.0] - 2017-04-21
+
+Updated with new properties for Moonstone controls.
+
 ## [1.0.0] - 2017-03-31
 
 ### Added